+++ tests/contrib/django/taskutils/backends/test_dummy.py
@@ -3,6 +3,7 @@ from typing import Any
 from uwsgiconf.contrib.django.uwsgify.taskutils.backends import BackendBase
 from uwsgiconf.contrib.django.uwsgify.taskutils.context import TaskContext
 from uwsgiconf.contrib.django.uwsgify.taskutils.decorators import task
+from uwsgiconf.runtime.locking import Lock
 from uwsgiconf.runtime.scheduling import register_cron
 
 
@@ -29,3 +30,20 @@ class MyContext(TaskContext):
 def test_with_ctx():
     task_1 = task(backend=BackendBase(context=MyContext))(mytask_ctx)
     assert task_1() == "{'1': 2}"
+
+
+def test_locked():
+    lock = Lock(1)
+
+    @register_cron()
+    @task(lock_skip=lock)
+    def task_1():
+        return 'done'
+
+    assert task_1() == "done"
+
+    lock.acquire()
+    assert task_1() is None
+
+    lock.release()
+    assert task_1() == "done"